FinalProject/services/mqtt_proxy_service/requirements.txt
blinker==1.9.0
cachetools==6.1.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
ecdsa==0.19.1
Flask==3.1.1
flask-cors==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
paho-mqtt==2.1.0
pyasn1==0.6.1
python-dotenv==1.1.0
python-jose==3.5.0
requests==2.32.4
rsa==4.9.1
six==1.17.0
urllib3==2.5.0
Werkzeug==3.1.3
gunicorn==21.2.0